requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
django-dotenv==1.4.2
djangorestframework==3.13.1
twilio==7.8.1
ipdb==0.13.9
djangorestframework-simplejwt==5.1.0
django-otp==1.1.3
pyotp==2.6.0
django-lint==2.0.4
django-cors-headers==3.11.0
cryptography==36.0.2
